// workspaces/kiali/plugins/kiali-react/src/types/IstioObjects.ts
export interface K8sMetadata {
  annotations?: { [key: string]: string };
  clusterName?: string;
  creationTimestamp?: string;
  deletionGracePeriodSeconds?: number;
  deletionTimestamp?: string;
  finalizers?: string[];
  generateName?: string;
  generation?: number;
  initializers?: K8sInitializers[];
  labels?: { [key: string]: string };
  name: string;
  namespace?: string;
  ownerReferences?: K8sOwnerReference[];
  resourceVersion?: string;
  selfLink?: string;
  uid?: string;
}

export interface K8sResource extends TypeMeta {
  metadata: K8sMetadata;
}

export interface IstioObject extends TypeMeta {
  metadata: K8sMetadata;
  status?: IstioStatus;
}

export interface IstioStatus {
  conditions?: StatusCondition[];
  validationMessages?: ValidationMessage[];
}
// ...
export interface PeerAuthentication extends IstioObject {
  spec: PeerAuthenticationSpec;
}

export interface PeerAuthenticationSpec {
  mtls?: PeerAuthenticationMutualTls;
  portLevelMtls?: { [key: number]: PeerAuthenticationMutualTls };
  selector?: PeerAuthenticationWorkloadSelector;
}

export interface PeerAuthenticationWorkloadSelector {
  matchLabels: { [key: string]: string };
}

export interface PeerAuthenticationMutualTls {
  mode: PeerAuthenticationMutualTLSMode;
}